+++ src/events.rs
@@ -1,0 +1,4 @@
+// Module declarations.
+pub(crate) mod event;
+pub(crate) mod event_detection;
+pub(crate) mod methods;