+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
-## 1.1.2
+## 0.2.0
+
+1. Added event detection.
+1. Renamed `IntegrationMethod` trait to `Integrator`. TODO actually do this
+
+## 0.1.2
 
 1. Added remaining Runge-Kutta integration methods.